ladp/requirements.txt
Flask
Flask-Mail
Flask-Redis
Jinja2
MarkupSafe
Pygments
Werkzeug
argparse
backports.shutil-get-terminal-size
blinker
click
decorator
enum34
ipdb
ipython
ipython-genutils
itsdangerous
pathlib2
pexpect
pickleshare
prompt-toolkit
ptyprocess
python-ldap
redis
scandir
simplegeneric
six
traitlets
wcwidth
wsgiref
